refactor: remove calorie rounding — all values Double with 1 decimal
- CalorieTarget: bmr/maintenance/target Int→Double
- CalorieCalculator: remove 'round to nearest 10', all returns Double
- Format helper fmt() shows whole numbers without .0, 1 decimal otherwise
- Cascade update: AppState, ViewModelProtocols, MealPlan, InsightsService,
  MigrationService, HeroSection, WeekCard, DayDetailCard, InsightsView
- Build number bump: 4→5

+++ Foodiary/Models/CalorieTarget.swift
@@ -4,9 +4,9 @@ import SwiftData
 @Model
 final class CalorieTarget {
     var id: UUID = UUID()
-    var bmr: Int = 0
-    var maintenanceCalories: Int = 0
-    var targetCalories: Int = 0
+    var bmr: Double = 0
+    var maintenanceCalories: Double = 0
+    var targetCalories: Double = 0
     var createdAt: Date = Date()
     var updatedAt: Date = Date()
     
@@ -14,9 +14,9 @@ final class CalorieTarget {
     
     init(
         id: UUID = UUID(),
-        bmr: Int,
-        maintenanceCalories: Int,
-        targetCalories: Int,
+        bmr: Double,
+        maintenanceCalories: Double,
+        targetCalories: Double,
         createdAt: Date = Date(),
         updatedAt: Date = Date()
     ) {
